hide quant_linear api
